+++ foundation/internal.py
@@ -18,12 +18,17 @@
 
 from tornado.web import *
 from tornado.gen import *
+from tornado.httpclient import *
+from collections import OrderedDict
 from foundation.place import PlacesOfInterest, slug_validation, visitor_only
+import json
+import os
+import time
 
 
 class PublicArea(PlacesOfInterest):
     @coroutine
-    @authentiated
+    @authenticated
     def post(self):
         public_path = os.path.join(self.settings["static_path"], "public")
         if self.request.files:
@@ -34,3 +39,104 @@ class PublicArea(PlacesOfInterest):
                 filename = f["filename"]
                 with open("configs/basic.json", "w+") as f:
                     return json.loads(f.read().replace(r"\n", "").strip())
+
+
+class AvatarArea(PlacesOfInterest):
+    @coroutine
+    @slug_validation(["hash"])
+    def get(self, slug):
+        size = self.get_arg("s", default=80, arg_type="number")
+        default = self.get_arg("d", default=404, arg_type="hash")
+        path = (self.settings["static_path"] + "/public/avatar/" + str(size) +
+                "/" + slug)
+        if os.path.exists(path):
+            with open(path + "/information.json") as f:
+                avatar_info = json.loads(f.read().replace(r"\n", "").strip())
+            if (int(time.time()) - avatar_info["time"]) <= (15 * 24 * 60 * 60):
+                self.set_header(
+                    "content-type", avatar_info["content_type"])
+                with open(path + "/content", "rb") as f:
+                    self.finish(f.read())
+                return
+            else:
+                os.removedirs(path)
+                os.makedirs(path)
+        else:
+            os.makedirs(path)
+        client = AsyncHTTPClient()
+        link = (
+            "https://secure.gravatar.com/avatar/" + slug + "?s=" +
+            str(size) + "&d=" + str(default))
+        response = yield client.fetch(link)
+        if response.error:
+            raise HTTPError(response.code)
+        avatar = response.body
+        content_type = response.headers.get("content-type")
+        avatar_info = {}
+        avatar_info["content_type"] = content_type
+        avatar_info["time"] = int(time.time())
+        with open(path + "/information.json", "w") as f:
+            f.write(json.dumps(avatar_info))
+        with open(path + "/content", "wb") as f:
+            f.write(avatar)
+        self.set_header("content-type", content_type)
+        self.finish(avatar)
+
+
+class ReplyArea(PlacesOfInterest):
+    @coroutine
+    def post(self):
+        action = self.get_arg("action", arg_type="hash")
+        writing_id = self.get_arg("writing", arg_type="number")
+        reply_id = self.get_arg("reply", arg_type="number")
+        if action == "get":
+            method = self.get_arg("method", arg_type="hash")
+            if method == "list" and writing_id:
+                result = yield self.get_reply(writing_id=writing_id)
+                for key in result:
+                    result[key]["emailmd5"] = self.hash(
+                        result[key]["email"].lower(), "md5")
+                    result[key]["email"] = None
+                    result[key]["ip"] = None
+            elif method == "single" and reply_id:
+                result = yield self.get_reply(id=reply_id)
+                result["emailmd5"] = self.hash(result["email"].lower(),
+                                               "md5", b64=False)
+                result["email"] = None
+                result["ip"] = None
+            else:
+                raise HTTPError(500)
+            self.finish(json.dumps(result))
+        elif action == "new":
+            reply = OrderedDict()
+            reply["writing_id"] = writing_id
+            if not self.current_user:
+                reply["master"] = False
+                reply["name"] = self.get_arg("name", arg_type="origin")
+                reply["email"] = self.get_arg("email", arg_type="mail_address")
+                reply["homepage"] = self.get_arg("homepage", arg_type="link")
+                reply["permit"] = False
+            else:
+                reply["master"] = True
+                reply["name"] = self.current_user["username"]
+                reply["email"] = self.current_user["email"]
+                reply["homepage"] = self.current_user["homepage"]
+                reply["permit"] = True
+            reply["ip"] = self.remote_ip
+            reply["time"] = int(time.time())
+            content = self.escape(self.get_arg("content", arg_type="origin"),
+                                  item_type="html")
+            reply["content"] = content
+            reply["_id"] = yield self.issue_id("Replies")
+            book = self.memories.select("Replies")
+            book.add(reply)
+            result = {}
+            try:
+                yield book.do()
+                result["success"] = True
+                result["id"] = reply["_id"]
+            except:
+                result["success"] = False
+            self.finish(json.dumps(result))
+        else:
+            raise HTTPError(500)