+++ spirit/nutrition/blossom/framework.js
@@ -46,6 +46,28 @@ function unixToDatetime(unix){
     return now.format("yyyy-MM-dd hh:mm:ss");
 }
 
+function buildWindow(){
+    if ($(window).width() < 750){
+        $("#reply-textarea-wrapper").css("width", "100%");
+        $("#reply-id-input").css("width", "calc(100% - 150px)");
+    }else{
+        $("#reply-textarea-wrapper").css("width", "calc(100% - 370px)");
+        $("#reply-id-input").css("width", "200px");
+    }
+}
+
+function getCookie(name) {
+    var r = document.cookie.match("\\b" + name + "=([^;]*)\\b");
+    return r ? r[1] : undefined;
+}
+
+jQuery.postJSON = function(url, args, callback) {
+    args._xsrf = getCookie("_xsrf");
+    $.ajax({url: url, data: $.param(args), dataType: "text", type: "POST",
+        success: function(response) {
+            callback(response);
+    }});
+};
 $(document).ready(function (){
     $(".change-time").html(function (){
         if ($(this).html() == "0"){
@@ -56,18 +78,109 @@ $(document).ready(function (){
     });
 });
 
-$(".textarea textarea").keypress(function(){
+$(".textarea textarea").keypress(function (){
     $(this).parent(".textarea").children("div").html($(this).val());
 });
-$(".textarea textarea").keydown(function(){
+$(".textarea textarea").keydown(function (){
     $(this).parent(".textarea").children("div").html($(this).val());
 });
-$(".textarea textarea").keyup(function(){
+$(".textarea textarea").keyup(function (){
     $(this).parent(".textarea").children("div").html($(this).val());
 });
-$(".textarea textarea").change(function(){
+$(".textarea textarea").change(function (){
     $(this).parent(".textarea").children("div").html($(this).val());
 });
-$(".textarea textarea").blur(function(){
+$(".textarea textarea").blur(function (){
     $(this).parent(".textarea").children("div").html($(this).val());
+});
+
+function getReplyData(){
+    result = {};
+    result.writing = $("#reply-writing").val();
+    result.name = $("#reply-name").val();
+    result.email = $("#reply-email").val();
+    result.homepage = $("#reply-homepage").val();
+    if ($("#reply-textarea").val().length <= 10){
+        throw("一个好的评论的长度应该大于10个字，难道不是么？");
+    }
+    result.content = $("#reply-textarea").val();
+    result.action = "new";
+    return result;
+}
+
+function buildReplyArea(){
+    replyData = {};
+    replyData.action = "get";
+    replyData.method = "list";
+    replyData.writing = window.writing_id;
+    $.postJSON("/channel/reply",
+        replyData,
+        function(data){
+            jsonData = JSON.parse(data);
+            $.each(jsonData, function (key, item){
+                $("#reply-list").prepend(function (){
+                    result = "<div class=\"reply-block\">";
+                    result += "<div class=\"reply-avatar\" style=\"background-image: url(/channel/avatar/" +
+                    item.emailmd5 + "?s=200&d=mm);\"></div><div class=\"reply-name\">" + item.name + "</div><div class=\"reply-time change-time\">" + unixToDatetime(Math.round(item.time)) + "</div><div class=\"reply-body\">" + item.content + "</div>";
+                    result += "</div>";
+                    return result;
+                });
+            });
+        }
+    );
+}
+
+function showNewReply(id){
+    replyData = {};
+    replyData.action = "get";
+    replyData.method = "single";
+    replyData.reply = id;
+    $.postJSON("/channel/reply",
+        replyData,
+        function(data){
+            $("#reply-list").prepend(function (){
+                jsonData = JSON.parse(data);
+                result = "<div class=\"reply-block\">";
+                result += "<div class=\"reply-avatar\" style=\"background-image: url(/channel/avatar/" +
+                jsonData.emailmd5 + "?s=200&d=mm);\"></div><div class=\"reply-name\">" + jsonData.name + "</div><div class=\"reply-time change-time\">" + unixToDatetime(Math.round(jsonData.time)) + "</div><div class=\"reply-body\">" + jsonData.content + "</div>";
+                result += "</div>";
+                return result;
+            });
+        }
+    );
+}
+
+function handleError(e){
+    $("#reply-alert").html(e);
+    $("#reply-alert").show();
+}
+
+function clearCurrentReply(){
+    $("#reply-textarea").val("");
+}
+
+$("#publish-reply").click(function (){
+    $("#reply-alert").hide();
+    try{
+    replyData = getReplyData();
+    if (replyData !== false){
+        try{
+            $.postJSON("/channel/reply",
+                replyData,
+                function(data){
+                    result = JSON.parse(data);
+                    if (result.success){
+                        showNewReply(result.id);
+                        clearCurrentReply();
+                    }else{
+                        throw("233");
+                    }
+                }
+            );
+       }catch(e){
+           throw("发生了未知错误，请稍候再试。");
+       }
+   }}catch(e){
+       handleError(e);
+   }
 });